move eventfilters under kcp with dual cluster mode
